kill CaseClassReaderPiece
